fix: correct gstreamer timestamp offset calculation for image header timestamp

+++ include/gscam/gscam.hpp
@@ -73,7 +73,7 @@ private:
 
   // ROS Inteface
   // Calibration between ros::Time and gst timestamps
-  double time_offset_;
+  uint64_t time_offset_;
   camera_info_manager::CameraInfoManager camera_info_manager_;
   image_transport::CameraPublisher camera_pub_;
   // Case of a jpeg only publisher
